+++ lib/nehm/configure.rb
@@ -2,20 +2,35 @@
 module Configure
   def self.menu
     loop do
-      puts 'Download path: ' + Paint[PathControl.dl_path, :magenta] if PathControl.dl_path
-      puts 'iTunes path: ' + Paint[PathControl.itunes_path_name, :magenta] if PathControl.itunes_path
-      puts 'Permalink: ' + Paint[Config[:permalink], :cyan] if Config[:permalink]
+      output = ''
+      options = [{ value: Config[:dl_path], name: 'Download path', color: :magenta },
+                 { value: Config[:permalink], name: 'Permalink', color: :cyan },
+                 { value: PathControl.itunes_root_path, name: 'iTunes path', color: :magenta },
+                 { value: PlaylistControl.playlist, name: 'iTunes playlist', color: :cyan }]
+
+      options.each do |option|
+        output << option[:name] + ': '
+        output <<
+          if option[:value] && !option[:value].to_s.empty?
+            Paint[option[:value], option[:color]]
+          else
+            Paint["doesn't set up", 'gold']
+          end
+        output << "\n"
+      end
+
+      puts output
       puts "\n"
 
       HighLine.new.choose do |menu|
         menu.prompt = Paint['Choose setting', :yellow]
 
         menu.choice('Edit download path'.freeze) { PathControl.set_dl_path }
-        menu.choice('Edit itunes path'.freeze) { PathControl.set_itunes_path } unless OS.linux?
         menu.choice('Edit permalink'.freeze) { UserControl.log_in }
-        menu.choice('Exit'.freeze) { exit }
+        menu.choice('Edit iTunes path'.freeze) { PathControl.set_itunes_path } unless OS.linux?
+        menu.choice('Edit iTunes playlist'.freeze) { PlaylistControl.set_playlist } unless OS.linux?
+        menu.choice('Exit'.freeze) { puts 'Goodbye!'; exit }
       end
-      sleep(1)
       puts "\n"
     end
   end